New and improved rotation functions using "delta swaps"!
Thanks to Jasper Neumann's generator:
http://programming.sirrida.de/bit_perm.html
http://programming.sirrida.de/calcperm.php

Rotation around the z-axis is basically permuting each 16 bits
in parallel, so it can be computed using the online generator.
Rotations around the other two axes require downloading the
generator's source code (now included in this repo) and changing
it to `#include "perm_b64.h"` instead of `"perm_b32.h"`, so that
it can handle 64-bit permutations. However, in retrospect, the
permutations for these rotations are basically isomorphic in terms
of their structure; you just have to modify the shifts and masks.